git-svn-id: svn://192.168.1.10/public/CoffeeMud@539 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Properties/Property.java
@@ -46,7 +46,7 @@ public class Property implements Ability, Cloneable
 
 	public boolean qualifies(MOB student){return false;}
 	public int qualifyingLevel(MOB student){return -1;}
-	public void startTickDown(Environmental affected, long tickTime){}
+	public void startTickDown(Environmental affected, int tickTime){}
 
 	public int profficiency(){return 0;}
 	public void setProfficiency(int newProfficiency){}
